third_party/WebKit/LayoutTests/inspector/elements/elements-panel-selection-after-delete.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function prepareTestTree()
{
    var template = document.querySelector("#testTree");
    var testTreeContainer = document.querySelector("#testTreeContainer");
    testTreeContainer.textContent = "";
    testTreeContainer.appendChild(document.importNode(template.content, true));
}

function test()
{
    function selectNode(className, callback)
    {
        var selector = "#testTreeContainer ." + className;
        InspectorTest.querySelector(selector, gotNode);

        function gotNode(node)
        {
            InspectorTest.selectNode(node).then(callback);
        }
    }

    function nodeToString(node)
    {
        if (!node)
            return "null"
        var result = "<" + node.nodeName();
        var attributes = node.attributes();
        for (var attribute of attributes) {
            result += " " + attribute.name;
            if (attribute.value)
                result += "=\"" + attribute.value + "\"";
        }
        result += ">";
        return result;
    }

    function prepareTestTree(callback)
    {
        InspectorTest.evaluateInPage("prepareTestTree()", callback);
    }

    function removeElementAsUser(element, callback)
    {
        InspectorTest.addSniffer(Elements.ElementsTreeOutline.prototype, "_updateModifiedNodes", callback);
        element.remove();
    }

    function removeElementExternally(element, callback)
    {
        var node = element.node();
        InspectorTest.addSniffer(Elements.ElementsTreeOutline.prototype, "_updateChildren", callback);
        node.removeNode();
    }

    function dumpSelectedNode()
    {
        var selectedElement = InspectorTest.firstElementsTreeOutline().selectedTreeElement;
        var selectedNode = selectedElement ? selectedElement.node() : null;
        InspectorTest.addResult("The following node is now selected: " + nodeToString(selectedNode));
    }

    InspectorTest.runTestSuite([
        function testUserDelete(next)
        {
            prepareTestTree(step2);

            function step2()
            {
                InspectorTest.addResult("Selecting node...");
                selectNode("child2", step3);
            }

            function step3()
            {
                dumpSelectedNode();

                InspectorTest.addResult("Deleting selected node like it was a user action...");
                var selectedElement = InspectorTest.firstElementsTreeOutline().selectedTreeElement;
                removeElementAsUser(selectedElement, step4);
            }

            function step4()
            {
                dumpSelectedNode();

                InspectorTest.addResult("Deleting selected node like it was a user action again...");
                var selectedElement = InspectorTest.firstElementsTreeOutline().selectedTreeElement;
                removeElementAsUser(selectedElement, step5);
            }

            function step5()
            {
                dumpSelectedNode();

                InspectorTest.addResult("Deleting last child like it was a user action...");
                var selectedElement = InspectorTest.firstElementsTreeOutline().selectedTreeElement;
                removeElementAsUser(selectedElement, step6);
            }

            function step6()
            {
                dumpSelectedNode();
                next();
            }
        },

        function testExternalDelete(next)
        {
            // We should wait for container node to be updated since it is already populated.
            InspectorTest.addSniffer(Elements.ElementsTreeOutline.prototype, "_updateModifiedNodes", step2);
            prepareTestTree();

            function step2()
            {
                InspectorTest.addResult("Selecting node...");
                selectNode("child2", step3);
            }

            function step3()
            {
                dumpSelectedNode();

                InspectorTest.addResult("Deleting selected node externally...");
                var selectedElement = InspectorTest.firstElementsTreeOutline().selectedTreeElement;
                removeElementExternally(selectedElement, step4);
            }

            function step4()
            {
                dumpSelectedNode();

                InspectorTest.addResult("Deleting selected node externally again...");
                var selectedElement = InspectorTest.firstElementsTreeOutline().selectedTreeElement;
                removeElementExternally(selectedElement, step5);
            }

            function step5()
            {
                dumpSelectedNode();

                InspectorTest.addResult("Deleting last child externally...");
                var selectedElement = InspectorTest.firstElementsTreeOutline().selectedTreeElement;
                removeElementExternally(selectedElement, step6);
            }

            function step6()
            {
                dumpSelectedNode();
                next();
            }
        },
    ]);
}

</script>
</head>
<body onload="runTest()">
<p>Tests that elements panel correctly updates selection on node removal.</p>
<template id="testTree">
    <div class="left">
        <div class="child1">
        </div>
        <div class="child2">
            <div class="child5">
            </div>
            <div class="child6">
            </div>
            <div class="child7">
            </div>
            <div class="child8">
            </div>
        </div>
        <div class="child3">
        </div>
    </div>
</template>
<div id="testTreeContainer">
</div>
</body>
</html>
