third_party/WebKit/LayoutTests/inspector/elements/elements-panel-selection-after-delete-expected.txt

Tests that elements panel correctly updates selection on node removal.


Running: testUserDelete
Selecting child3 node...
Deleting selected node like it was a user action...
The following node is now selected: <DIV class="child4">
Deleting selected node like it was a user action again...
The following node is now selected: <DIV class="left">

Running: testExternalDelete
Selecting child3 node...
Deleting selected node externally...
The following node is now selected: <DIV class="left">
Deleting selected node externally again...
The following node is now selected: <DIV id="testTreeContainer">
